fixes the list view so it can handle much more complex scenarios
