+++ src/components/BeagleContainer/styled.ts
@@ -19,6 +19,5 @@ import styled from 'styled-components'
 export const StyledContainer = styled.div`
 	display: flex;
   flex-direction: column;
-  overflow: auto;
   flex-wrap: wrap;
 `
